+++ multiqc/modules/trim_galore/__init__.py
@@ -1,0 +1,3 @@
+from .trim_galore import MultiqcModule
+
+__all__ = ["MultiqcModule"]